Moved the merging of electrodes to the DNPU instead of letting them on the processor
